requirements.txt
python-telegram-bot==20.8
playwright==1.41.2
google-genai
nest-asyncio
